# docs/openapi/password-management.yaml
openapi: 3.0.3
info:
  title: Synaxis Password Management API
  description: Enterprise-grade password management endpoints with policy enforcement
  version: 1.0.0
servers:
  - url: https://api.synaxis.com/v1
    description: Production server
  - url: https://staging-api.synaxis.com/v1
    description: Staging server
  - url: http://localhost:8080/v1
    description: Development server

tags:
  - name: Password Management
    description: User password operations
  - name: Password Policy
    description: Organization password policy management

paths:
  /users/me/password:
    post:
      tags:
        - Password Management
      summary: Change password
      description: |
        Changes the current user's password with policy enforcement.
        Validates the new password against the organization's password policy,
        checks password history to prevent reuse, and updates the password expiration date.
      operationId: changePassword
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
            examples:
              valid:
                summary: Valid password change request
                value:
                  currentPassword: "OldPassword123!"
                  newPassword: "NewSecurePassword456!"
      responses:
        '200':
          description: Password changed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangePasswordResponse'
        '400':
          description: Bad request - invalid password or validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                weakPassword:
                  summary: Password too weak
                  value:
                    success: false
                    errorMessage: "Password must be at least 12 characters long"
                reusedPassword:
                  summary: Password reused
                  value:
                    success: false
                    errorMessage: "You cannot reuse your last 5 passwords"
                incorrectCurrentPassword:
                  summary: Current password incorrect
                  value:
                    success: false
                    errorMessage: "Current password is incorrect"
        '401':
          description: Unauthorized - authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '423':
          description: Locked - too many failed attempts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                errorMessage: "Too many failed attempts. Password change locked for 15 minutes"

  /users/me/password/validate:
    post:
      tags:
        - Password Management
      summary: Validate password
      description: |
        Validates a password against the organization's password policy
        without changing the password. Returns validation results and strength score.
      operationId: validatePassword
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidatePasswordRequest'
            examples:
              strong:
                summary: Strong password
                value:
                  password: "VeryStrongPassword!@#123ABC"
              weak:
                summary: Weak password
                value:
                  password: "weak"
      responses:
        '200':
          description: Password validation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PasswordValidationResponse'
        '401':
          description: Unauthorized - authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /organizations/{organizationId}/password-policy:
    get:
      tags:
        - Password Policy
      summary: Get password policy
      description: |
        Retrieves the password policy for the specified organization.
        Returns default policy if none has been configured.
      operationId: getPasswordPolicy
      security:
        - BearerAuth: []
      parameters:
        - name: organizationId
          in: path
          required: true
          description: The organization ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Password policy retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PasswordPolicy'
        '401':
          description: Unauthorized - authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - user is not a member of the organization
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Organization not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Password Policy
      summary: Update password policy
      description: |
        Updates the password policy for the specified organization.
        Requires organization admin or owner role.
      operationId: updatePasswordPolicy
      security:
        - BearerAuth: []
      parameters:
        - name: organizationId
          in: path
          required: true
          description: The organization ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePasswordPolicyRequest'
            examples:
              strict:
                summary: Strict password policy
                value:
                  minLength: 16
                  requireUppercase: true
                  requireLowercase: true
                  requireNumbers: true
                  requireSpecialCharacters: true
                  passwordHistoryCount: 10
                  passwordExpirationDays: 60
                  passwordExpirationWarningDays: 7
                  maxFailedChangeAttempts: 3
                  lockoutDurationMinutes: 30
                  blockCommonPasswords: true
                  blockUserInfoInPassword: true
              relaxed:
                summary: Relaxed password policy
                value:
                  minLength: 8
                  requireUppercase: false
                  requireLowercase: true
                  requireNumbers: true
                  requireSpecialCharacters: false
                  passwordHistoryCount: 3
                  passwordExpirationDays: 0
                  passwordExpirationWarningDays: 0
                  maxFailedChangeAttempts: 5
                  lockoutDurationMinutes: 15
                  blockCommonPasswords: true
                  blockUserInfoInPassword: false
      responses:
        '200':
          description: Password policy updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PasswordPolicy'
        '400':
          description: Bad request - invalid policy configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidMinLength:
                  summary: Minimum length too short
                  value:
                    errorMessage: "MinLength must be between 8 and 128"
        '401':
          description: Unauthorized - authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - user is not an organization admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Organization not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT authentication token

  schemas:
    ChangePasswordRequest:
      type: object
      required:
        - currentPassword
        - newPassword
      properties:
        currentPassword:
          type: string
          format: password
          minLength: 1
          description: The current password
          example: "OldPassword123!"
        newPassword:
          type: string
          format: password
          minLength: 1
          description: The new password
          example: "NewSecurePassword456!"

    ChangePasswordResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Whether the password change was successful
          example: true
        passwordExpiresAt:
          type: string
          format: date-time
          nullable: true
          description: The password expiration date (null if never expires)
          example: "2026-05-12T13:48:53Z"

    ValidatePasswordRequest:
      type: object
      required:
        - password
      properties:
        password:
          type: string
          format: password
          minLength: 1
          description: The password to validate
          example: "MySecurePassword123!"

    PasswordValidationResponse:
      type: object
      properties:
        isValid:
          type: boolean
          description: Whether the password is valid according to policy
          example: true
        strengthScore:
          type: integer
          minimum: 0
          maximum: 100
          description: Password strength score (0-100)
          example: 85
        strengthLevel:
          type: string
          enum: [Very Weak, Weak, Fair, Good, Strong]
          description: Password strength level
          example: "Strong"
        errors:
          type: array
          description: Validation errors (empty if valid)
          items:
            type: string
          example: []
        warnings:
          type: array
          description: Validation warnings (e.g., weak password suggestions)
          items:
            type: string
          example: []

    PasswordPolicy:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Password policy ID
          example: "550e8400-e29b-41d4-a716-446655440000"
        organizationId:
          type: string
          format: uuid
          description: Organization ID
          example: "550e8400-e29b-41d4-a716-446655440001"
        minLength:
          type: integer
          minimum: 8
          maximum: 128
          description: Minimum password length
          example: 12
        requireUppercase:
          type: boolean
          description: Require uppercase letters
          example: true
        requireLowercase:
          type: boolean
          description: Require lowercase letters
          example: true
        requireNumbers:
          type: boolean
          description: Require numbers
          example: true
        requireSpecialCharacters:
          type: boolean
          description: Require special characters
          example: true
        passwordHistoryCount:
          type: integer
          minimum: 0
          maximum: 24
          description: Number of previous passwords to remember
          example: 5
        passwordExpirationDays:
          type: integer
          minimum: 0
          maximum: 365
          description: Password expiration period in days (0 = never expires)
          example: 90
        passwordExpirationWarningDays:
          type: integer
          minimum: 0
          maximum: 30
          description: Days before expiration to show warning
          example: 14
        maxFailedChangeAttempts:
          type: integer
          minimum: 3
          maximum: 10
          description: Maximum failed password change attempts before lockout
          example: 5
        lockoutDurationMinutes:
          type: integer
          minimum: 5
          maximum: 60
          description: Lockout duration in minutes after failed attempts
          example: 15
        blockCommonPasswords:
          type: boolean
          description: Block common passwords
          example: true
        blockUserInfoInPassword:
          type: boolean
          description: Block user info (email, name) in password
          example: true
        createdAt:
          type: string
          format: date-time
          description: Policy creation timestamp
          example: "2026-02-11T12:48:53Z"
        updatedAt:
          type: string
          format: date-time
          description: Policy last update timestamp
          example: "2026-02-11T12:48:53Z"

    UpdatePasswordPolicyRequest:
      type: object
      properties:
        minLength:
          type: integer
          minimum: 8
          maximum: 128
          default: 12
          description: Minimum password length
        requireUppercase:
          type: boolean
          default: true
          description: Require uppercase letters
        requireLowercase:
          type: boolean
          default: true
          description: Require lowercase letters
        requireNumbers:
          type: boolean
          default: true
          description: Require numbers
        requireSpecialCharacters:
          type: boolean
          default: true
          description: Require special characters
        passwordHistoryCount:
          type: integer
          minimum: 0
          maximum: 24
          default: 5
          description: Number of previous passwords to remember
        passwordExpirationDays:
          type: integer
          minimum: 0
          maximum: 365
          default: 90
          description: Password expiration period in days (0 = never expires)
        passwordExpirationWarningDays:
          type: integer
          minimum: 0
          maximum: 30
          default: 14
          description: Days before expiration to show warning
        maxFailedChangeAttempts:
          type: integer
          minimum: 3
          maximum: 10
          default: 5
          description: Maximum failed password change attempts before lockout
        lockoutDurationMinutes:
          type: integer
          minimum: 5
          maximum: 60
          default: 15
          description: Lockout duration in minutes after failed attempts
        blockCommonPasswords:
          type: boolean
          default: true
          description: Block common passwords
        blockUserInfoInPassword:
          type: boolean
          default: true
          description: Block user info (email, name) in password

    ErrorResponse:
      type: object
      properties:
        errorMessage:
          type: string
          description: Error message describing what went wrong
          example: "Password must be at least 12 characters long"
        success:
          type: boolean
          description: Whether the operation was successful
          example: false